nic_api/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
pytesseract==0.3.13
python-bidi==0.6.6
scikit-image==0.25.2
scipy==1.15.2